Switched to Decimals

+++ Deribit/Core/Messages/MarketData/GetInstrumentsMessage.cs
@@ -18,21 +18,21 @@ namespace Deribit.Core.Messages.MarketData
     public class GetInstrumentsResponse : IResponse<GetInstrumentsResponse[]>
     {
         public string base_currency { get; set; }
-        public float block_trade_commission { get; set; }
-        public float contract_size { get; set; }
+        public decimal block_trade_commission { get; set; }
+        public decimal contract_size { get; set; }
         public long creation_timestamp { get; set; }
         public long expiration_timestamp { get; set; }
         public string instrument_name { get; set; }
         public bool is_active { get; set; }
         public string kind { get; set; }
         public int? leverage { get; set; }
-        public float maker_commission { get; set; }
-        public float? min_trade_amount { get; set; }
+        public decimal maker_commission { get; set; }
+        public decimal? min_trade_amount { get; set; }
         public string? option_type { get; set; }
         public string quote_currency { get; set; }
         public string settlement_period { get; set; }
-        public float? strike { get; set; }
-        public float taker_commission { get; set; }
-        public float tick_size { get; set; }
+        public decimal? strike { get; set; }
+        public decimal taker_commission { get; set; }
+        public decimal tick_size { get; set; }
     }
 }
